Move BannerTitle to Markup
- Move `BannerTitle` to `Markup`
- Delete `BannerTitle` component folder
- Update imports for `Home` and `UserDashboard`
