--- src/components/BannerTitle/BannerTitle.styled.tsx
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-export const BannerTitleContainer = styled.div``;
-
-export const Title = styled.h1`
-  font-family: "Trebuchet MS";
-`;
